linera_persistent::file

Struct File

pub struct File<T> { /* private fields */ }

An implementation of Persist based on an atomically-updated file at a given path. An exclusive lock is taken using flock(2) to ensure that concurrent updates cannot happen, and writes are saved to a staging file before being moved over the old file, an operation that is atomic on all Unixes.

Implementations

impl<T: Serialize + DeserializeOwned> File<T>

pub fn new(path: &Path, value: T) -> Result<Self, Error>

Creates a new persistent file at path containing value.

pub fn read(path: &Path) -> Result<Self, Error>

Reads the value from a file at path, returning an error if it does not exist.

pub fn read_or_create( path: &Path, value: impl FnOnce() -> Result<T, Error>, ) -> Result<Self, Error>

Reads the value from a file at path, calling the value function to create it if it does not exist. If it does exist, value will not be called.

pub fn save(&self) -> Result<(), Error>